restaurant-recommendation/phase-2-recommendation-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0
